feat: Se agregaron las validaciones a el CRUD de Empleados

+++ RepuestosAutomotrices/src/main/java/com/marcosmontenegro/service/EmpleadoServiceImplements.java
@@ -28,29 +28,61 @@ public class EmpleadoServiceImplements implements EmpleadoService{
 
     @Override
     public Empleado saveEmpleado(Empleado empleado) throws RuntimeException {
-        return empleadoRepository.save(empleado);
+        try {
+            if (empleado == null || empleado.getNombreEmpleado().isBlank()|| empleado.getApellidoEmpleado().isBlank()|| empleado.getPuestoEmpleado().isBlank() || empleado.getEmailEmpleado().isBlank()){
+                throw new IllegalArgumentException("Todos Los campos son obligatorios");
+            }
+
+            if (!empleado.getEmailEmpleado().contains("@gmail.com") || empleado.getEmailEmpleado().contains("@yahoo.com")
+                    || empleado.getEmailEmpleado().contains("@outlook.com") || empleado.getEmailEmpleado().contains("@icloud.com") || empleado.getEmailEmpleado().contains("@hotmail.com")){
+                throw new IllegalArgumentException("El correo electronico debe tener un dominio como @gmail.com, @yahoo.com, @outlook.com, @icloud.com y @hotmail.com");
+            }
+
+            if (empleadoRepository.existsByNombreProveedorAndTelefonoProveedorAndDireccionProveedorAndEmailProveedor(
+                    empleado.getNombreEmpleado() ,
+                    empleado.getApellidoEmpleado(),
+                    empleado.getPuestoEmpleado(),
+                    empleado.getEmailEmpleado()
+            )){
+                throw new RuntimeException("El empleado ya existe");
+            }
+            return empleadoRepository.save(empleado);
+        }catch (Exception e) {
+            throw new RuntimeException(e.getMessage());
+        }
     }
+
 
 
     @Override
     public Empleado updateEmpleado(Integer id, Empleado empleado){
-        Empleado existingEmpleado = empleadoRepository.findById(id).orElseThrow(() -> new RuntimeException("El empleado no existe"));
 
-            existingEmpleado.setNombreEmpleado(empleado.getNombreEmpleado());
-            existingEmpleado.setApellidoEmpleado(empleado.getApellidoEmpleado());
-            existingEmpleado.setPuestoEmpleado(empleado.getPuestoEmpleado());
-            existingEmpleado.setEmailEmpleado(empleado.getEmailEmpleado());
-
-            return empleadoRepository.save(existingEmpleado);
+            try {
+                if (!empleado.getEmailEmpleado().contains("@gmail.com") || empleado.getEmailEmpleado().contains("@yahoo.com")
+                    || empleado.getEmailEmpleado().contains("@outlook.com") || empleado.getEmailEmpleado().contains("@icloud.com") || empleado.getEmailEmpleado().contains("@hotmail.com")){
+                throw new IllegalArgumentException("El correo electronico debe tener un dominio como @gmail.com, @yahoo.com, @outlook.com, @icloud.com y @hotmail.com");
+            }
+            return empleadoRepository.save(empleado);
+            } catch (RuntimeException e){
+            throw new RuntimeException(e.getMessage());
+        }
     
     }
 
 
     @Override
     public void deleteEmpleado(Integer id) {
-        if (!empleadoRepository.existsById(id)) {
-         throw new RuntimeException("Empleado no existe");
+        try {
+            if (id == null || id == 0){
+                throw new RuntimeException("El id esta vacío o es igual a 0");
+            }
+
+            if (!empleadoRepository.existsById(id)) {
+                throw new RuntimeException("El id no existe");
+            }
+            empleadoRepository.deleteById(id);
+        }catch (Exception e){
+            throw new RuntimeException(e.getMessage());
         }
-        empleadoRepository.deleteById(id);
     }
 }
